requirements.txt
autopep8==1.4.4
Click==7.0
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
imutils==0.5.2
numpy==1.16.2
opencv-python==4.1.0.25
Pillow==6.0.0
psycopg2==2.8.2
pycodestyle==2.5.0
scikit-learn==0.20.3
scipy==1.2.1
